Don't try to update chai higher than 4.x.x.
Updating chai seems to cause the ts checks to fail. Based on [this
comment](https://github.com/chaijs/chai/issues/1568#issuecomment-1873883627)
and others in an issue thread in the Chai project repo, version 5.0.0
and higher of chai requires a ceratin type of module use. Perhaps this
is the reason why the tests fail for us -- I can't tell without
looking more closely. Right now, this is not an essential update to
do, and we need to stop Dependabot trying to update this every time it
runs.