Add missing license files & do minor updates to dot files (#7161)
* Add missing license headers

Apparently I misread guidance somewhere, and license headers are
required for these sorts of files after all.

* Revise to address comments in review

This removes unnecessary or unused content, in response to review
comments on PR #7161.

* Change copyright to refer to Cirq Developers

* Add back max_line_length setting

After more discussion, the conclusion is that despite the issues with
`max_line_length`, it is handled as expected in the editors that many
people have been using.

* Change copyright to refer to Cirq Developers

* Add temp max_line_length value for TS linter

The typescript linter looks at max_line_length and takes it into
account if it finds a value. This is causing CI checks to fail. We
need to reformat the sources in question and then we can remove this
setting again

* Shorten the file header

---------

Co-authored-by: Pavol Juhas <juhas@google.com>