Update `_NPY_MAXDIMS` (#7167)
* Update `_NPY_MAXDIMS` 

numpy 1.x supported a maximum of 32 dimensions and then raised an error .. this was an issue for large scale simulations with qubits up to 35 for state vector simulation and up to 18 qubits for density matrix simulations.

As a workaround to support these cases we worked with 1D and 2D arrays for these cases which was very inefficient. 

now numpy 2.0 supports up to 64 dimensions

* use the value from numpy instead of hardcoding it

* use suggestion from @maffoo

* Update cirq-core/cirq/linalg/transformations.py

Co-authored-by: Pavol Juhas <juhas@google.com>

* nit

* Update transformations.py

* Handle negative dimensions in can_numpy_support_shape

And add a direct unit test.

---------

Co-authored-by: Pavol Juhas <juhas@google.com>