Add additional #how-to-cite anchor
I noticed that several of the other projects have (or had) sections
titled How to Cite, and then in some documentation files, the links to
citation info uses `#how-to-cite` anchors. Just to be safe in case not
all links have been updated everywhere, let's add an additional
hyperlink anchor `#how-to-cite` here too.