Update and make minor additions to docs/ in time for Cirq 1.5 (#7175)
* Add citation for Tesseract Decoder

* Add Tesseract Decoder & remove Rigetti

* Remove statement about cirq-rigetti

It seems like the Docker container may not be used solely for the
Rigetti module, so I left the other text about Docker and only removed
the statement involving Rigetti.

* Add mention of Tesseract & fix a couple of tiny niggles

* Remove mention of labels no longer used

This entire section still needs a good re-think and more revisions.
For now, I just removed mention of the labels that have been changed.

* Repeat the statement about vendor modules

It's true that the section on what's covered mentioned that vendor
modules are not, but it seemed strange that it was not mentioned in
the section that is specifically about what's no covered.

* Edit for typos

Mostly fix capitalization and occasional missing hyphens.

* Add notices that cirq-rigetti is deprecated

* Small edits for grammar and clarity

* Move paragraph about hardware modules higher in the list

Putting the bullet item about hardware modules seems higher in the
list seems like it's better, so that the flow of concepts goes
somewhare more general to somewhat more specific.

* Trivial edits to rewrite a phrase & add formatting

* Fix trivial typos

* Partially rewrite support.md

Changes:

* There were 3 H1 headings, and at the same time, the first heading
  did not match what is in the TOC. I changed this to have one H1
  heading that matches the TOC text, made the remaining headings H2,
  and added a lead-in paragaraph

* Rewrote the section about using the issue tracker in attempt to be
  more specific and clear.

* Added a section about using the Q.C. Stack Exchange.

* Added a section about using email to contact us.

* Fix minor inconsistent heading content style

A trivial change to match the way other headings were written.

* Update triage description

Removed more no-longer-existing issue labels, and rewrote section on
stale issue/PR heading to be more succinct.

* Fix trivial typos

Capitalization and placement of colon characters.

* Run spell checker

* Fix a few more typos

* Fix a few more tiny typos

* Take out Tesseract for now

It has no Cirq interface, and it's a bit too early stage still.

* Update list of software

* Spell out LBL and ORNL

* Mildly edit another description

* Fix misspelled name

* Remove cirq-rigetti from list of example modules

It's not essential, and removing it now will help reduce the number of
places we have to update when we remove Rigetti altogether.

* Make note of deprecation status in more places

* Update the section on Python imports

This adds information about `isort`'s use in Cirq, now that PR #7181
has been merged and `isort` is used in the check/ scripts.

* Fix typo

* Shorten the import ordering section

---------

Co-authored-by: Pavol Juhas <juhas@google.com>