Add PR labeler workflow with size labeling (for now)
This size labeling workflow should be a close approximation to what
CirqBot currently does via `pr_monitor`. Once in place, it should be
possible to remove pr_monitor from the Cirq code base, because the
automerge functionality of pr_monitor (which is the other thing that
it did) is now implemented using GitHub's merge queues.

This basically resolves #7000.