Invoke the labeler program directly
There is no way to get the error messages from the labeler action. We
need to get the output if we want to note that something failed
without failing the whole workflow. Thankfully, the action can be
invoked more directly using `podman`, and it's not hard to revise the
workflow to capture the output and write it as an annotation on the
summary page.