Use a different action for size labeler GHA (#7219)
* Use a different action for size labeler GHA

The previous labeler was buggy and poorly written.

* Need set permissions for id-token

* Update .github/pr-size-labels.yaml

Co-authored-by: Pavol Juhas <pavol.juhas@gmail.com>

* Widen the scope of excluded package-lock.json files

* Improve explanation of how the settings work

* Don't fail the workflow if labeling fails

* Invoke the labeler program directly

There is no way to get the error messages from the labeler action. We
need to get the output if we want to note that something failed
without failing the whole workflow. Thankfully, the action can be
invoked more directly using `podman`, and it's not hard to revise the
workflow to capture the output and write it as an annotation on the
summary page.

* Clarify comments some more

---------

Co-authored-by: Pavol Juhas <pavol.juhas@gmail.com>