Reorder cirq + cirq-rigetti module install & import
Pasting the following code into a cell in a fresh notebook in
colab.research.google.com,

```python
try:
    import cirq
    import cirq_rigetti
except ImportError:
    print("installing cirq-rigetti...")
    !pip install --quiet cirq-rigetti
    print("installed cirq-rigetti.")
    import cirq
    import cirq_rigetti
```

produces the error described in issue #7267. The reason is the
following. The `cirq-rigetti` module requires NumPy version 1.x, so
pip-installing the Rigetti module causes NumPy 1.x to be installed.
However, the Colab environment comes with version 2.x preloaded. When
Cirq is imported, it results in an import of `numpy`, which (because the
NumPy module outside the Python environment has been changed) results in
NumPy 1.x being imported into a Python runtime that already has NumPy 2
loaded.

Changing the order a little bit seems to work: pip-install Cirq first,
import it (which doesn't result in errors because it doesn't load a
conflicting version of NumPy), then after that, pip-install cirq-rigetti
separately, followed by importing it.

This allows the packages to load. Unfortunately, I can't get the
forest-sdk package to install in my Colab environment, so I can't go far
in testing whether the rest of the notebook works.