Improve handling of arguments & provide help
This addresses review comments about how the PR number is provided; this
new version lets the user provide it via command line argument or
environment variable.

In addition, the GitHub token and repository name can be provided on the
command line as well.

Finally, this also offers a `--help` option and prints usage information.