Replace PR labeler with better & safer version (#7271)
* Remove old size labeler config file

No longer needed because we're switching to a different tool.

* Start a subdirectory for tools used in CI workflows

Although there is only one script here at this time, I've often felt
some of the actions I've written would have been better implemented as
scripts. It may be time to start a subdirectory where such things are
placed.

* Add label-pr-size.sh to dev_tools/

This program was originally written by @lovesegfault (Bernardo Meurer)
for Quantum AI's Pyle around the year 2022.

* Replace PR labeler with better & safer versions

This replaces the previous action with the use of a script
ported from Pyle.

* Add timeout to job

* Update dev_tools/ci/size-labeler.sh

Co-authored-by: Pavol Juhas <pavol.juhas@gmail.com>

* Update .github/workflows/pr-labeler.yaml

Co-authored-by: Pavol Juhas <pavol.juhas@gmail.com>

* Update dev_tools/ci/size-labeler.sh

Co-authored-by: Pavol Juhas <pavol.juhas@gmail.com>

* Add "*.bundle.js"

* Improve handling of arguments & provide help

This addresses review comments about how the PR number is provided; this
new version lets the user provide it via command line argument or
environment variable.

In addition, the GitHub token and repository name can be provided on the
command line as well.

Finally, this also offers a `--help` option and prints usage information.

* Mark LABELS and IGNORED as read-only

* Take advantage of Bash shift argument

* Check for missing arg values

* Use awk to prefix stderr output from jq

`jq_file` has code to return its status.  Do not exit the script there.

* Remove BASH_VERSION check

The script is intended for GHA which have compatible bash.

* Simplify size-labeler to get its inputs from environment only

Remove unnecessary argument processing.  The script is for
automated execution and is not supposed to be run by users.

* Use PR_NUMBER environment variable in the pr-labeler workflow

Match environment variable name expected by size-labeler.sh.

---------

Co-authored-by: Pavol Juhas <pavol.juhas@gmail.com>
Co-authored-by: Pavol Juhas <juhas@google.com>