Optimize the HHL algorithm using amplitude amplification (#7141)
* Optimize the HHL algorithm using amplitude amplification.

I changed the example to simulate two circuits: the first one is the
previous HHL circuit in the example without amplitude amplification,
and the second one is the HHL circuit with amplitude amplification.

The example demonstrates that amplitude amplification obtains an estimate
faster (using fewer total calls to simulate.run).

* Add an explanation for the number of iterations in amplitude amplification.

* Add parameter types and factor the output print to a separate method.

* Delete empty line.