Use list for GoogleNoiseProperties.readout_errors values (#7392)
Problem: The values in
`SuperconductingQubitsNoiseProperties.readout_errors`
and `GoogleNoiseProperties.readout_errors` dictionaries are declared as
`np.ndarray` in docstrings, but annotated as`list[float]` in the code,
and can be set to either type. JSON serialization fails when assigned a
list.

Solution: Always convert to `list[float]` following the type annotation.

This fixes serialization failure for `GoogleNoiseProperties` objects
created
with `noise_properties_from_calibration`.