Use an efficient representation for connected components of operations in the merge methods.
Currently merge_operations represents merged components as a CircuitOperation.
This means in a merge of n operations, n-1 CircuitOperations are created,
with a complexity of O(n^2).
We use a disjont-set data structure to reduce the complexity to O(n) for
merge_k_qubit_unitaries_to_circuit_op. merge_operations itself can't be
improved because it uses a merge_func that requires creation of
CircuitOperation at every step.