Make the points in cirq_google v2 proto to be double (#7498)
In order to make it forward and backward compatible, i added an extra
field `points_fl64` and
1. serialization phase: always write to both `float points` and `double
points_fl64`
2. deserialization phase: use `points_fl64` first if it presented, if
not, fall back to `points`.

It will temporary make the communication cost slightly higher since we
communicating two points. But it will be make the transition much
easier.

UPDATE: added the double field for LinSpace and Const as well.

---------

Co-authored-by: Pavol Juhas <pavol.juhas@gmail.com>