Adjust example notebooks for the change of the use_repetion_ids default to false (#7504)
Adjust the `docs/build/classical_control.ipynb` notebook for the change
of
CircuitOperation default to `use_repetition_ids=False` in #7265.
This restores the notebook output as it was before the default change.

Other notebooks that were showing FutureWarning for the new default were
not affected by rendering with the old and new setting.  These were:

- docs/google/best_practices.ipynb
- docs/hardware/pasqal/getting_started.ipynb
- docs/simulate/virtual_engine_interface.ipynb
- docs/start/intro.ipynb
- docs/transform/routing_transformer.ipynb

Related to #7232