Don't attempt to make the logo light/dark responsive (#7512)
PyPI does not implement a dark mode, and due to the fact that SVG's
inside `<img>` tags cannot access properties outside of themselves, it
does not work to try to switch colors on PyPI.

The least problematic solution seems to be to go back to using a single
color for the text of the logo, with a change to use a lighter shade of
gray so that it's more legible on GitHub in dark mode. This shade of
gray is not the official Cirq logo shade of gray; it's a compromise
intended to make the text work on both white and black backgrounds.

This commit also changes the names of the styles to be more meaningful,
because at some point someone might want to revisit this thing, and this
may help them figure out what's what in this SVG file.