Fix global phase computation in matrix gate decomposition (#7482)
Fixes #7473

Old code with `cirq.unitary(circuit)` performs the unitary assuming the
qubit order is ascending based on the qubit id. (The identity operator
in that code helped when some qubits were factored out during
decomposition, but didn't help with ordering). Changed to
`circuit.unitary(qubits, qubits)` to make the qubit order explicit. The
first one defines the ordering to be the same as the decomposed
matrixgate, and the second makes sure that all qubits are retained even
if the decomposition factored some out)