Fix sidepanel contents for section on noise (#7518)
The removal of some documents in a previous PR left the left-side TOC
panel with an extra heading. Now, it also turns out that none of the
other sections in the Cirq docs use an accordion, so we can feed two
birds with one scone by getting rid of the accordion and moving the
contents up one level. This should result in a properly-structured
panel.

<div align="center">
<img width="50%" alt="image"
src="https://github.com/user-attachments/assets/fd90fb7e-4085-448f-8e09-016df4348212"
/>
 
</div>