Regenerate QuantumEngineService gRPC client libraries with GAPIC (#7500)
This PR contains changes made by re-running the
[GAPIC](https://googleapis.github.io/gapic-generators/) code generators
for the QuantumEngineService, using the Google-internal proto
definitions as input. Generated changes include:

- REST-based RPC transport in addition to gRPC
- A new RPC method `QuantumEngineService.GetQuantumProcessorConfig`,
along with the associated request and response objects
- Field renaming `whitelisted_users` -> `allowlisted_users`
- Updated type hints

In addition to the automated tests, I also verified that the updated RPC
clients work by creating a colab which used the updated `cirq_google`
library to send jobs to the production QuantumEngineService.

---------

Co-authored-by: Pavol Juhas <juhas@google.com>