Add option and set the default float encoding as float32 (#7554)
Adding an option in the `run_context_to_proto` so that only one of the
float32 or float64 will be written into proto. Default to be float 32 so
that the behavior is unchanged.

When we update the Cirq version internally to include this CL, we need
to make sure all swithcing into `v2.run_context_to_proto(..,
use_float64=True)`