Clarify the instructions about pre-release notebooks (#7573)
The previous instructions in `docs/dev/notebooks.md` said to do the
steps "at release time", but that was ambiguous (IMHO) and also
contradicted the instructions in
`dev_tools/notebooks/isolated_notebook_tests.py`. This is an attempt to
clarify the instructions and make them consistent between the two files.

---------

Co-authored-by: Pavol Juhas <juhas@google.com>