Performance improvement in _check_duplicate_keys (#7599)
- This change improves the performance of _check_duplicate_keys.
- Rather than check each sweep individually, this adds the keys of all
the sweeps to the set and checks the total size. The total size of the
set should be the sum of the sizes of the sweep, otherwise there are
duplicates.
- This is a relatively minor performance improvement, but I saw it while
trying to optimize something else, and it is self-contained.

This speeds checking a large sweep (the example I have is a Zip of 700k
sweeps) from 300ms to 150ms.