Delegate SingleQubitPauliStringGateOperation mul, rmul to its PauliString base
SingleQubitPauliStringGateOperation inherits from GateOperation and
PauliString in that resolution order, i.e., with `__mul__`, `__rmul__`
defaulting to those of GateOperation.  To avoid infinite recursion
we make it use the PauliString operations instead.

Also avoid creating a new operation in `_try_interpret_as_pauli_string`
if the argument is SingleQubitPauliStringGateOperation which is already
a PauliString.