Fixing ts-build errors and build config from webpack-to-vite merge (#7697)
Fixes an issue found in #7687 where artifacts were not properly building
with Vite. The appropriate CI output can be found
[here](https://github.com/quantumlib/Cirq/actions/runs/18333406290/job/52212831218#step:5:1).
See discussion in original PR for more details.

The issue stems from the fact that Vite does not allow handling multiple
entry points from an array in configs, which is different than Webpack.
A common alternative is to use separate config files for each entry
point, but this PR leverages [Vite's `--mode`
flag](https://vite.dev/guide/env-and-mode) to allow bundling from one
file. Should the number of widgets in `cirq-web` grow, it may be worth
looking at consolidating the widgets in one entry point, but that seems
like overkill for now.

The large line count is due to how Vite bundles the TS, but the file
size is still comparable to the old bundle files (501KB versus 503KB).