Remove import of typing_extensions
As of #7590 typing_extensions is not in cirq-core requirements.
mypy can detect mismatched `merge` arguments w/r to the base class
without the `@typing_extensions.override` decorator.  It does not
seem worth reintroducing this dependency, here we drop the decorator
instead.

NB: Python 3.12 added a similar `typing.override` decorator,
but we need to support 3.11.