Add OSSF Scorecard configuration file
This configuration file tells the scanner to ignore unpinned
dependencies in the CI workflows. They are a known concern and are
judged to be relatively low risk for the time being. They will be
addressed in future work. Until then, it is not useful to keep seeing
the same warnings repeatedly.