pr-size-labeler - fetch change info for the maximum number of files
By default the GHA endpoint `pulls/PULL_NUMBER/files` retrieves
change data for up to 30 files which is insufficient for PRs that
touch more files.  Here we increase the count to the maximum
per_page limit of 100.

It does not seem worthwhile to add pagination processing as
in the last year only 1% of PRs touched more than 100 files.

Ref: https://docs.github.com/en/rest/pulls/pulls?apiVersion=2022-11-28#list-pull-requests-files