Fix CircuitDag constructor to match its networkx.DiGraph base (#7771)
Problem: `CircuitDag` had a different number (2 vs 1) and a different
order of positional arguments than its `networkx.DiGraph` base.
This caused failure for networkx-3.6 which added `DiGraph.__new__`
as it was receiving incompatible arguments.

Solution: Change `CircuitDag` arguments so they are compatible with
its superclass, i.e., make `incoming_graph_data` the first argument
as in the base and change `can_reorder` to a keyword-only argument.

This breaks the `CircuitDag` interface, but that should be acceptable
for a `contrib` function.