Adjust workflow job runner choices for better efficiency (#7775)
After confirming with higher levels of our org that we have the budget
for it, I configured customized larger GitHub-hosted runners for the
Quantumlib organization. These new runners are 8-core runners for Linux
and Windows. (GitHub offers even larger runners, but let's see first how
much using 8-cores will speed up the workflows in case there are some
other bottlenecks that prevent us from gaining any further advantages.)

By contrast, some of the workflows can also make use of `ubuntu-slim`
runners, which are faster to start and cheaper to run. As I understand
it, those slim runners have only 1 core, so I looked for jobs that don't
do things that can be parallelized. The slim runners take longer to run
the jobs, but they're priced at 1/8 the cost per minute of the regular
runners, and the particular jobs they're being used for are not
bottlenecks; thus, the longer execution of these small jobs doesn't
affect the overall workflow time.

GitHub offers a macos-15-large runner with 12 cores. They're about 10x
the price of regular runners, so we want to apply them carefully.

Finally, I found that the `pytest -n auto` flag does not make full use
of the available vCPUS. It seems to use only 1/2 of the total available.
So to make full use of the runners, I changed to `-n logical` to use the
number of processors available.

Demonstration of the speed improvements: first, a run from yesterday
without the changes:

Total duration: 12m 18s

Now a run (this PR) with the changes:

Total duration: 9m 2s

---------

Co-authored-by: Pavol Juhas <juhas@google.com>