Fix inconsistency between qubits measurements in readout circuits and real measurement circuits  (#7801)
There is an inconsistency on qubits measurements in readout circuits and
measurement circuits. The measurement circuits always measure the full
circuit set of qubits, while the qubits being measured in
readout/calibration circuits are based only on the qubits present in the
Pauli strings.
In this PR, I have updated the circuit generation to only measure the
qubits relevant to the Pauli group being processed.